workers/requirements.txt
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pillow==10.2.0
face_recognition==1.3.0
imagehash==4.3.1
numpy==1.26.3
python-magic==0.4.27
exifread==3.0.0
opencv-python-headless==4.9.0.80
requests==2.31.0
flower==2.0.1
minio==7.2.3
